otaku-rest-api/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==8.0.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
dnspython==2.2.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
jikanpy==4.3.2
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==6.0.2
pymongo==4.2.0
python-dotenv==0.20.0
pytz==2020.1
requests==2.28.1
simplejson==3.17.6
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.11
Werkzeug==1.0.1
yarl==1.8.1
